requirements.txt
fastapi
uvicorn[standard]
pyjwt
cryptography
